requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==11.3.0
psycopg2-binary==2.9.10
setuptools
gunicorn==21.2.0
whitenoise==6.6.0

aiohttp
requests